#21: refactor structure
