+++ pkg/api/providers/openai/chat.go
